requirements.txt
python-dotenv
beautifulsoup4
requests
mysql-connector-python
sqlalchemy
pymysql